requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-common==1.1.28
azure-core==1.38.1
azure-identity==1.25.2
azure-mgmt-authorization==4.0.0
azure-mgmt-core==1.6.0
azure-mgmt-managementgroups==1.1.0
azure-mgmt-resource==25.0.0
azure-mgmt-subscription==3.1.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
frozenlist==1.8.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
isodate==0.7.2
jinja2==3.1.6
librt==0.8.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
microsoft-kiota-abstractions==1.9.8
microsoft-kiota-authentication-azure==1.9.8
microsoft-kiota-http==1.9.8
microsoft-kiota-serialization-form==1.9.8
microsoft-kiota-serialization-json==1.9.8
microsoft-kiota-serialization-multipart==1.9.8
microsoft-kiota-serialization-text==1.9.8
msal==1.34.0
msal-extensions==1.3.1
msgraph-core==1.3.8
msgraph-sdk==1.54.0
msrest==0.7.1
multidict==6.7.1
mypy==1.19.1
mypy-extensions==1.1.0
oauthlib==3.3.1
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pathspec==1.0.4
pluggy==1.6.0
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
pyjwt==2.11.0
pytest==9.0.2
pytest-asyncio==1.3.0
pyyaml==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.2
ruff==0.15.1
shellingham==1.5.4
std-uritemplate==2.0.8
typer==0.23.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
yarl==1.22.0
zipp==3.23.0
